add redesign message
Calsen currenly, contains a lot of repeated code and a lot of bad design choices.
Which has made adding features tedious. And also the fact the some of the utils
functions may not be stable have lead to the idea of a complete redesign in a better
language for this sort of project.